feat: basic tailwind outline
