+++ client/src/pages/ActivityPage.tsx
@@ -1,11 +1,7 @@
 import { Link } from 'react-router-dom';
-import ActivityList from '../components/ActivityList';
-
-
+import ActivityList from '../components/Activity/ActivityList';
 
 function ActivityPage() {
-
-  
   return (
     <>
       <ActivityList />
