+++ client/src/pages/TripPage.tsx
@@ -4,13 +4,13 @@ import { useEffect } from 'react';
 import { useDispatch } from 'react-redux';
 import { AppDispatch } from '../app/store';
 import { fetchUserTrips } from '../services/fetchTrip';
-import TripComponent from '../components/TripComponent';
+import TripComponent from '../components/Trip/TripComponent';
 
 function TripPage() {
   const dispatch = useDispatch<AppDispatch>();
 
-  //hardcoding user_id until we create actual functionality 
-  const user_id = 1; 
+  //hardcoding user_id until we create actual functionality
+  const user_id = 1;
 
   useEffect(() => {
     dispatch(fetchUserTrips(user_id));
